# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /chat/messages:
    get:
      operationId: listChatMessages
      tags: [chat]
      summary: List chat companion messages
      description: Returns all chat messages for the current session
      responses:
        "200":
          description: List of chat messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChatMessage"
    post:
      operationId: sendChatMessage
      tags: [chat]
      summary: Send a message to the companion
      description: Sends a message to the AI companion and returns user message and bot response
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendChatMessageInput"
      responses:
        "200":
          description: Message and companion response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SendChatMessageResponse"
        "400":
          description: Invalid chat input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
      required: [error]
    ChatMessage:
      type: object
      properties:
        id:
          type: string
        sessionId:
          type: string
        role:
          type: string
          enum: [user, assistant]
        content:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - sessionId
        - role
        - content
        - createdAt
    SendChatMessageInput:
      type: object
      properties:
        message:
          type: string
          minLength: 1
          maxLength: 2000
      required:
        - message
    SendChatMessageResponse:
      type: object
      properties:
        userMessage:
          $ref: "#/components/schemas/ChatMessage"
        botMessage:
          $ref: "#/components/schemas/ChatMessage"
      required:
        - userMessage
        - botMessage
...